list all documents
Resolved all issues
