--- apps/citadel/list_all_documents/__init__.py
@@ -1,8 +1,0 @@
-# -*- encoding: utf-8 -*-
-"""
-Copyright (c) 2023 - Aark Global
-"""
-
-from flask import Blueprint
-
-blueprint = Blueprint("list_all_documents_blueprint", __name__, url_prefix="")
