+++ apps/common/constants.py
@@ -30,9 +30,9 @@ MAX_PARALLEL_FILE_UPLOADS_ALLOWED = 21
 MAX_ALLOWED_FILE_UPLOADS_ON_ONE_PAGE = 150
 
 # Default_folders_name
-DEFAULT_INCOMING_SUBFOLDER = "/Incoming-files/"
-DEFAULT_VALIDATION_SUCCESSFUL_SUBFOLDER = "/Validation Successful/"
-DEFAULT_VALIDATION_FAILED_SUBFOLDER = "/Validation Failed/"
+DEFAULT_INCOMING_SUBFOLDER = "/Incoming-Files/"
+DEFAULT_VALIDATION_SUCCESSFUL_SUBFOLDER = "/Validation-Successful/"
+DEFAULT_VALIDATION_FAILED_SUBFOLDER = "/Validation-Failed/"
 DEFAULT_INPROGRESS_SUBFOLDER = "/Inprogress/"
 DEFAULT_SUCCESSFUL_SUBFOLDER = "/Successful/"
 DEFAULT_FAILED_SUBFOLDER = "/Failed/"
