+++ apps/services/company_folder_creator.py
@@ -7,7 +7,7 @@ def create_company_folder_structure(pk: str):
     container_client = utils.get_azure_storage_blob_container_client(constants.DEFAULT_BLOB_CONTAINER)
     # Create the container
     if not container_client.exists():
-        logging.info(
+        logging.error(
             "%s container doesnot exist, creating container %s",
             constants.DEFAULT_BLOB_CONTAINER,
             constants.DEFAULT_BLOB_CONTAINER,
@@ -27,6 +27,7 @@ def create_company_folder_structure(pk: str):
         create_folder(f"{constants.COMPANY_ROOT_FOLDER_PREFIX}{pk}", subfolder, container_client)
 
 
+# creating dummy file inside every company's sub-folder
 def create_folder(company_folder: str, subfolder: str, container_client: ContainerClient):
     blob_client = container_client.get_blob_client(f"{company_folder}{subfolder}dummy")
     blob_client.upload_blob(b"")
